restructured frontend folder structure
