+++ db-scheduler-ui-frontend/src/components/common/RefreshCircle.tsx
