+++ db-scheduler-ui-frontend/src/components/common/TitleRow.tsx
@@ -15,7 +15,7 @@ import React from 'react';
 
 import { Box, HStack } from '@chakra-ui/react';
 
-import { SortButton } from 'src/components/SortButton';
+import { SortButton } from 'src/components/input/SortButton';
 import colors from 'src/styles/colors';
 import { SortBy } from 'src/models/QueryParams';
