+++ db-scheduler-ui-frontend/src/components/input/FilterBox.tsx
