--- db-scheduler-ui-frontend/src/styles/components.ts
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) Bekk
- *
- * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
- *
- * <p>http://www.apache.org/licenses/LICENSE-2.0
- *
- * <p>Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-// Override or extend ChakraUI components here
-
-import { ComponentStyleConfig } from '@chakra-ui/react';
-
-const buttonStyles: ComponentStyleConfig = {
-  baseStyle: {
-    fontWeight: 'bold',
-  },
-  variants: {
-    primary: (props) => ({
-      bg: props.colorMode === 'dark' ? 'white' : 'primary.500',
-    }),
-  },
-};
-
-const components = {
-  Button: buttonStyles,
-};
-
-export default components;
