+++ db-scheduler-ui-frontend/src/styles/index.ts
@@ -12,16 +12,9 @@
  * limitations under the License.
  */
 import colors from './colors';
-import typography from './typography';
-import breakpoints from './breakpoints';
-import components from './components';
 
-// All styles are placeholders added as examples
 const theme = {
   colors,
-  typography,
-  breakpoints,
-  components,
 };
 
 export default theme;
