--- db-scheduler-ui-frontend/src/styles/typography.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) Bekk
- *
- * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
- *
- * <p>http://www.apache.org/licenses/LICENSE-2.0
- *
- * <p>Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-const typography = {
-  fonts: {
-    body: "'Open Sans', sans-serif",
-    heading: "'Roboto', sans-serif",
-  },
-  fontSizes: {
-    xs: '0.75rem',
-    sm: '0.875rem',
-    md: '1rem',
-    lg: '1.125rem',
-    xl: '1.25rem',
-    // ... and so on
-  },
-  // ... other typography settings
-};
-
-export default typography;
